backend/requirements.txt
fastapi
uvicorn
pymongo
pydantic
python-dotenv
httpx
eth-account
cryptography
